+++ src/backend/playlist-server/playlist-server/src/main/java/com/sgwannabe/playlistserver/playlist/util/KeyGenerator.java
@@ -3,7 +3,7 @@ package com.sgwannabe.playlistserver.playlist.util;
 public class KeyGenerator {
     private static final String PLAYLIST_KEY_PREFIX = "playlist";
 
-    public static String cartKeyGenerate(String id) {
+    public static String playlistKeyGenerate(String id) {
         return PLAYLIST_KEY_PREFIX + ":" + id;
     }
 }
